- Separate message frame detector class for each KV*Message is removed.
