--- Replicated_Storage_Service/src/weloveclouds/communication/utils/detector/KVAdminMessageFrameDetector.java
@@ -1,11 +1,0 @@
-package weloveclouds.communication.utils.detector;
-
-import weloveclouds.commons.serialization.models.XMLTokens;
-
-public class KVAdminMessageFrameDetector extends AbstractMessageFrameDetector {
-
-    public KVAdminMessageFrameDetector() {
-        super(XMLTokens.KVADMIN_MESSAGE);
-    }
-
-}
