--- Replicated_Storage_Service/src/weloveclouds/communication/utils/detector/KVMessageFrameDetector.java
@@ -1,11 +1,0 @@
-package weloveclouds.communication.utils.detector;
-
-import weloveclouds.commons.serialization.models.XMLTokens;
-
-public class KVMessageFrameDetector extends AbstractMessageFrameDetector {
-
-    public KVMessageFrameDetector() {
-        super(XMLTokens.KVMESSAGE);
-    }
-
-}
