--- Replicated_Storage_Service/src/weloveclouds/communication/utils/detector/KVTransactionMessageFrameDetector.java
@@ -1,11 +1,0 @@
-package weloveclouds.communication.utils.detector;
-
-import weloveclouds.commons.serialization.models.XMLTokens;
-
-public class KVTransactionMessageFrameDetector extends AbstractMessageFrameDetector {
-
-    public KVTransactionMessageFrameDetector() {
-        super(XMLTokens.KVTRANSACTION_MESSAGE);
-    }
-
-}
